分页查询参数统一

+++ project/src/main/java/com/nageoffer/shortlink/project/dto/req/RecyclePageDTO.java
@@ -23,10 +23,10 @@ public class RecyclePageDTO {
     /**
      * 页码
      */
-    private int current;
+    private Long current;
 
     /**
      * 每页大小
      */
-    private int size;
+    private Long size;
 }
